maestro/pyproject.toml
[tool.poetry]
name = "maestro"
version = "0.1.0"
description = ""
authors = ["ipark"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
Flask = "^3.0.3"
requests = "^2.32.3"
websockets = "^13.1"
logging = "^0.4.9.6"


[tool.poetry.group.dev.dependencies]
python-dotenv = "^1.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
